Add manage listings

+++ app/Http/Controllers/ListingController.php
@@ -47,6 +47,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully.');
@@ -61,6 +63,11 @@ class ListingController extends Controller
     // Update listing data
     public function update(Request $request, Listing $listing)
     {
+        //Confirm logged user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -83,7 +90,18 @@ class ListingController extends Controller
     // Delete listing
     public function delete(Listing $listing)
     {
+        //Confirm logged user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted');
+    }
+
+    // Manage listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
